requirements.txt
streamlit
langchain-google-genai
python-dotenv
pdfplumber
google-generativeai